Expose SafeEventEmitterProvider type
Not all of the types within `./src/utils/cache.ts` seemed to be solely
related to caching. These have been extracted to a separate file, and
the main file has been updated to export `SafeEventEmitterProvider` from
this file.
