+++ src/block-ref.ts
@@ -6,11 +6,8 @@ import {
 } from 'json-rpc-engine';
 import clone from 'clone';
 import pify from 'pify';
-import {
-  Block,
-  SafeEventEmitterProvider,
-  blockTagParamIndex,
-} from './utils/cache';
+import { blockTagParamIndex } from './utils/cache';
+import type { Block, SafeEventEmitterProvider } from './types';
 
 interface BlockRefMiddlewareOptions {
   blockTracker?: PollingBlockTracker;
